+++ TractorMod/release-notes.md
@@ -1,9 +1,9 @@
 # Release notes
-## 4.4 (upcoming)
+## 4.4
 * Added support for using the tractor in the mines.
 * Added options to toggle all standard features (thanks to Fox536!).
 * Fixed tractor appearing in social menu.
-* Fixed issue where tractor doesn't appear in Farm Expansion's carpenter menu (requires Farm Expansion 3.3+).
+* Fixed tractor not appearing in Farm Expansion's carpenter menu (requires Farm Expansion 3.3+).
 
 ## 4.3
 * Updated to SMAPI 2.4.